requirements.txt
streamlit
openai
elasticsearch
fastapi[all]
uvicorn
python-dotenv